Suspension overhaul
Biiig break after handing in dissertation, I'm so back. Overhauled suspension. Wheels now actually move instead of visually. A lot harder to fall through the floor, but curbs don't affect anywhere near enough now. Also a problem with one frame touching ground and the next not, evolves to the tire model not producing any force because it thinks it isn't touching the ground. Most noticeable in the lack of top speed.
Also changed the tire load to use the suspension force. Tire's have way to much grip now that I am working it out properly, but I can lower the grip value easily, keeping the same for now for consistency with suspension testing.