Added Raycast system
Theoretically works, don't actually know yet though. Preparing to overhaul the suspension overhaul. I was a bit naive with the last suspension change. In reality just having the tire detect collision isn't really enough. Kind of going back to how it was before but more proper. Raycasts are fairly standard in engines and they are of course useful for lots of things. Could have moved the ray collider to the suspension and done stuff like that but this way is a lot better. Hopefully next push is with the suspension system using a ray collider to detect collision.